+++ src/components/pages/page-favorites-empty/favorites-empty.tsx
