+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './components/app/App';
+import App from './components/app/app';
 import {Provider} from 'react-redux';
 import store from './store/store';
 import {checkAuth} from './store/api-actions';
